ADEV v4.91:1089 (r0.9.0)
Following the current engine archetecture for current version number. Main version number will increment with  the engine until source, we'll fork/branch at that point. Refer to the release number (r) for the current game version of Ashendell as well as it's progress.

This release is the first development upload, if you are working on the project or are a staff member please keep all of your files up to date before submitting your files for entry into the game, this will save a lot of time.

Notes:

Everything added is here is new to this rep, nothing to say really.